All projects self-manage their own DI concerns
