+++ Letterbook.Adapter.Db/Letterbook.Adapter.Db.csproj
@@ -38,7 +38,9 @@
         <PrivateAssets>all</PrivateAssets>
       </PackageReference>
       <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.0" />
+      <PackageReference Include="Npgsql.OpenTelemetry" Version="8.0.3" />
       <PackageReference Include="OpenIddict.EntityFrameworkCore" Version="4.7.0" />
+      <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.6.0" />
     </ItemGroup>
 
     <ItemGroup>
