+++ Letterbook.Api/Program.cs
@@ -1,13 +1,15 @@
 using Letterbook.Adapter.ActivityPub;
+using Letterbook.Adapter.Db;
 using Letterbook.Api.Authentication.HttpSignature;
 using Letterbook.Api.Authentication.HttpSignature.DependencyInjection;
 using Letterbook.Api.Swagger;
-using Letterbook.Config;
 using Letterbook.Core;
 using Letterbook.Core.Exceptions;
 using Letterbook.Core.Extensions;
+using Letterbook.Core.Models;
 using Letterbook.Workers;
 using MassTransit;
+using Microsoft.AspNetCore.Identity;
 using Serilog;
 using Serilog.Enrichers.Span;
 using Serilog.Events;
@@ -51,7 +53,10 @@ public class Program
 			;
 		builder.Services.AddActivityPubClient(builder.Configuration);
 		builder.Services.AddServices(builder.Configuration);
-		builder.Services.AddIdentity();
+		builder.Services.AddIdentity<Account, IdentityRole<Guid>>()
+			.AddEntityFrameworkStores<RelationalContext>()
+			.AddDefaultTokenProviders()
+			.AddDefaultUI();
 		builder.Services.AddMassTransit(bus => bus.AddWorkerBus(builder.Configuration));
 
 		builder.WebHost.UseUrls(coreOptions.BaseUri().ToString());
