--- Letterbook.Config/Letterbook.Config.csproj
@@ -1,36 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Letterbook.Config</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="MassTransit" Version="8.2.2" />
-      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.0" />
-      <PackageReference Include="Npgsql" Version="8.0.3" />
-      <PackageReference Include="Npgsql.OpenTelemetry" Version="8.0.3" />
-      <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.8.1" />
-      <PackageReference Include="OpenTelemetry.Exporter.Prometheus.AspNetCore" Version="1.6.0-rc.1" />
-      <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.6.0" />
-      <PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.8.1" />
-      <PackageReference Include="Serilog" Version="3.1.1" />
-      <PackageReference Include="Serilog.Enrichers.Span" Version="3.1.0" />
-      <PackageReference Include="Serilog.Expressions" Version="4.0.0" />
-      <PackageReference Include="Serilog.Sinks.Grafana.Loki" Version="8.3.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Reference Include="Npgsql.OpenTelemetry">
-        <HintPath>..\..\..\.nuget\packages\npgsql.opentelemetry\8.0.3\lib\netstandard2.0\Npgsql.OpenTelemetry.dll</HintPath>
-      </Reference>
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Letterbook.Adapter.Db\Letterbook.Adapter.Db.csproj" />
-      <ProjectReference Include="..\Letterbook.Core\Letterbook.Core.csproj" />
-    </ItemGroup>
-
-</Project>
