+++ Letterbook.sln
@@ -74,8 +74,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Api.Authenticati
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook", "Letterbook\Letterbook.csproj", "{11BCFCEF-2E6A-48DF-A6BD-37CC80E25EDB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Config", "Letterbook.Config\Letterbook.Config.csproj", "{C1B93B23-7229-4004-A99B-55CCF6BCC519}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Workers", "Letterbook.Workers\Letterbook.Workers.csproj", "{E53BF0A3-8A77-45A9-8CFF-606F20B46789}"
 EndProject
 Global
@@ -143,10 +141,6 @@ Global
 		{11BCFCEF-2E6A-48DF-A6BD-37CC80E25EDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{11BCFCEF-2E6A-48DF-A6BD-37CC80E25EDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{11BCFCEF-2E6A-48DF-A6BD-37CC80E25EDB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C1B93B23-7229-4004-A99B-55CCF6BCC519}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C1B93B23-7229-4004-A99B-55CCF6BCC519}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C1B93B23-7229-4004-A99B-55CCF6BCC519}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C1B93B23-7229-4004-A99B-55CCF6BCC519}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E53BF0A3-8A77-45A9-8CFF-606F20B46789}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E53BF0A3-8A77-45A9-8CFF-606F20B46789}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E53BF0A3-8A77-45A9-8CFF-606F20B46789}.Release|Any CPU.ActiveCfg = Release|Any CPU
